cpuset: option for forcing recursive cpus and mems update
Trying to set cpuset.cpus of a cgroup so that cpus of a child group
would not be a subset of parent cpus anymore causes an error (device
or resource busy). This patch adds an option to force the update
recursively to all child groups. "mems" is handled similarly to
"cpus".

Option --child-policy="force" sets cpuset.cpus of the group and all
its children to the same value. Introducing a string-valued option
instead of a bool leaves room for future policies. For instance, a
safer policy could modify only those childgroups that already had the
same value as their parent, or limit the scope to only cpus or mems.

+++ libcontainer/cgroups/fs/cpuset.go
@@ -4,6 +4,7 @@ package fs
 
 import (
 	"errors"
+	"fmt"
 	"os"
 	"path/filepath"
 	"strconv"
@@ -28,7 +29,11 @@ func (s *CpusetGroup) Apply(path string, d *cgroupData) error {
 
 func (s *CpusetGroup) Set(path string, r *configs.Resources) error {
 	if r.CpusetCpus != "" {
+		if r.ChildPolicy == "force" {
+			return fmt.Errorf("--child-policy force NOT IMPLEMENTED") // TODO: set recursively to all children
+		}
 		if err := cgroups.WriteFile(path, "cpuset.cpus", r.CpusetCpus); err != nil {
+
 			return err
 		}
 	}
